Extract common logic for getting resources
