+++ CommandLine/src/Commands/GetGame.cs
@@ -1,9 +1,6 @@
 using System;
-using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
-using Worms.Logging;
 using Worms.Resources;
 using Worms.Resources.Games;
 
@@ -16,8 +13,7 @@ namespace Worms.Commands
     [Command("game", "games", "replay", "replays", "WAgame", Description = "Retrieves information for Worms games (.WAgame files)")]
     internal class GetGame : CommandBase
     {
-        private readonly IResourceRetriever<GameResource> _gameRetriever;
-        private readonly IResourcePrinter<GameResource> _printer;
+        private readonly ResourceGetter<GameResource> _gameRetriever;
 
         [Argument(
             0,
@@ -26,10 +22,9 @@ namespace Worms.Commands
                 "Optional: The name or search pattern for the Game to be retrieved. Wildcards (*) are supported")]
         public string Name { get; }
 
-        public GetGame(IResourceRetriever<GameResource> gameRetriever, IResourcePrinter<GameResource> printer)
+        public GetGame(ResourceGetter<GameResource> gameRetriever)
         {
             _gameRetriever = gameRetriever;
-            _printer = printer;
         }
 
         public Task<int> OnExecuteAsync(IConsole console)
@@ -37,7 +32,7 @@ namespace Worms.Commands
             try
             {
                 var windowWidth = Console.WindowWidth == 0 ? 80 : Console.WindowWidth;
-                Print(Name, console.Out, windowWidth);
+                _gameRetriever.PrintResources(Name, console.Out, windowWidth);
             }
             catch (ConfigurationException exception)
             {
@@ -46,32 +41,6 @@ namespace Worms.Commands
             }
 
             return Task.FromResult(0);
-        }
-
-        private void Print(string name, TextWriter writer, int outputMaxWidth)
-        {
-            var requestForAll = string.IsNullOrWhiteSpace(name);
-            var userSpecifiedName = !requestForAll && !name.Contains('*');
-            var matches = requestForAll ? _gameRetriever.Get() : _gameRetriever.Get(name);
-
-            if (userSpecifiedName)
-            {
-                switch (matches.Count)
-                {
-                    case 0:
-                        throw new ConfigurationException($"No Game found with name: {name}");
-                    case 1:
-                        _printer.Print(writer, matches.Single(), outputMaxWidth);
-                        break;
-                    default:
-                        _printer.Print(writer, matches, outputMaxWidth);
-                        break;
-                }
-            }
-            else
-            {
-                _printer.Print(writer, matches, outputMaxWidth);
-            }
         }
     }
 }
