+++ CommandLine/src/Commands/GetScheme.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
 using Worms.Logging;
 using Worms.Resources;
+using Worms.Resources.Games;
 using Worms.Resources.Schemes;
 
 // ReSharper disable MemberCanBePrivate.Global - CLI library uses magic to read members
@@ -16,8 +17,7 @@ namespace Worms.Commands
     [Command("scheme", "schemes", "wsc", Description = "Retrieves information for Worms Schemes (.wsc files)")]
     internal class GetScheme : CommandBase
     {
-        private readonly IResourceRetriever<SchemeResource> _schemesRetriever;
-        private readonly IResourcePrinter<SchemeResource> _printer;
+        private readonly ResourceGetter<SchemeResource> _schemesRetriever;
 
         [Argument(
             0,
@@ -26,10 +26,9 @@ namespace Worms.Commands
                 "Optional: The name or search pattern for the Scheme to be retrieved. Wildcards (*) are supported")]
         public string Name { get; }
 
-        public GetScheme(IResourceRetriever<SchemeResource> schemesRetriever, IResourcePrinter<SchemeResource> printer)
+        public GetScheme(ResourceGetter<SchemeResource> schemesRetriever)
         {
             _schemesRetriever = schemesRetriever;
-            _printer = printer;
         }
 
         public Task<int> OnExecuteAsync(IConsole console)
@@ -37,7 +36,7 @@ namespace Worms.Commands
             try
             {
                 var windowWidth = Console.WindowWidth == 0 ? 80 : Console.WindowWidth;
-                PrintScheme(Name, console.Out, windowWidth);
+                _schemesRetriever.PrintResources(Name, console.Out, windowWidth);
             }
             catch (ConfigurationException exception)
             {
@@ -46,32 +45,6 @@ namespace Worms.Commands
             }
 
             return Task.FromResult(0);
-        }
-
-        private void PrintScheme(string name, TextWriter writer, int outputMaxWidth)
-        {
-            var requestForAll = string.IsNullOrWhiteSpace(name);
-            var userSpecifiedName = !requestForAll && !name.Contains('*');
-            var matches = requestForAll ? _schemesRetriever.Get() : _schemesRetriever.Get(name);
-
-            if (userSpecifiedName)
-            {
-                switch (matches.Count)
-                {
-                    case 0:
-                        throw new ConfigurationException($"No Scheme found with name: {name}");
-                    case 1:
-                        _printer.Print(writer, matches.Single(), outputMaxWidth);
-                        break;
-                    default:
-                        _printer.Print(writer, matches, outputMaxWidth);
-                        break;
-                }
-            }
-            else
-            {
-                _printer.Print(writer, matches, outputMaxWidth);
-            }
         }
     }
 }
